refactor(meeting_facilitator/session): split tests to sibling tests_session.rs (#1266)
- src/meeting_facilitator/session.rs: 610 → 332 LOC
- New src/meeting_facilitator/tests_session.rs (276 LOC)

Pure refactor: 19 tests pass.
